# neatly/docs/swagger.yaml
basePath: /
definitions:
  account.LoginAccountDTO:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  account.RegisterAccountDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  account.WithTokenDTO:
    properties:
      email:
        type: string
      name:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  e.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  note.CreateNoteDTO:
    properties:
      body:
        type: string
      color:
        type: string
      header:
        type: string
    required:
    - color
    - header
    type: object
  note.GetAllNotesDTO:
    properties:
      notes:
        items:
          $ref: '#/definitions/note.Note'
        type: array
    type: object
  note.Note:
    properties:
      body:
        type: string
      color:
        type: string
      edited:
        type: string
      header:
        type: string
      id:
        type: integer
      shortBody:
        type: string
      tags:
        description: '[]tag.Tag'
        items:
          $ref: '#/definitions/tag.Tag'
        type: array
    type: object
  note.UpdateNoteDTO:
    properties:
      body:
        type: string
      color:
        type: string
      header:
        type: string
      id:
        type: integer
    type: object
  tag.CreateTagDTO:
    properties:
      color:
        type: string
      name:
        type: string
    required:
    - color
    - name
    type: object
  tag.GetAllTagsDTO:
    properties:
      tags:
        items:
          $ref: '#/definitions/tag.Tag'
        type: array
    type: object
  tag.Tag:
    properties:
      color:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - color
    - name
    type: object
  tag.UpdateTagDTO:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for notes-taking applications
  termsOfService: http://swagger.io/terms/
  title: Neat.ly API
  version: "1.0"
paths:
  /api/v1/accounts/login:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/account.LoginAccountDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.WithTokenDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      summary: Login
      tags:
      - account
  /api/v1/accounts/register:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/account.RegisterAccountDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      summary: Register
      tags:
      - register
  /api/v1/notes:
    get:
      consumes:
      - application/json
      description: create note
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/note.GetAllNotesDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all notes from user
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: create note
      parameters:
      - description: note content
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/note.CreateNoteDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create note
      tags:
      - notes
  /api/v1/notes/{id}:
    delete:
      consumes:
      - application/json
      description: delete note
      operationId: delete-note
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Note
      tags:
      - notes
    get:
      consumes:
      - application/json
      description: get note by id
      operationId: get-note-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/note.Note'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Note By Id
      tags:
      - notes
    patch:
      consumes:
      - application/json
      description: update note
      operationId: update-note
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: note content
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/note.UpdateNoteDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Note
      tags:
      - notes
  /api/v1/notes/{id}/tags:
    get:
      consumes:
      - application/json
      description: get tags for note
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.GetAllTagsDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all tags on one note
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: create tag
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: tag info
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create tag
      tags:
      - tags
  /api/v1/notes/search:
    get:
      consumes:
      - application/json
      description: search note
      operationId: search-note
      parameters:
      - description: notes search by tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/note.GetAllNotesDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search Note
      tags:
      - search
  /api/v1/tags:
    get:
      consumes:
      - application/json
      description: get tags from user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.GetAllTagsDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all tags
      tags:
      - tags
  /api/v1/tags/{id}:
    delete:
      consumes:
      - application/json
      description: delete one tag by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete one tag by ID
      tags:
      - tags
    get:
      consumes:
      - application/json
      description: get one tag by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.Tag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get one tag by ID
      tags:
      - tags
    patch:
      consumes:
      - application/json
      description: update one tag by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: tag info
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/tag.UpdateTagDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update tag by ID
      tags:
      - tags
  /api/v1/tags/{id}/tags/{tag_id}:
    delete:
      consumes:
      - application/json
      description: detach tag by ID from note by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: tag id
        in: path
        name: tag_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/e.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/e.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Detach tag by ID from note by ID
      tags:
      - tags
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"